+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Added
 - added ability to use multiple env vars per line in yaml files
 - added ``priority`` property of policies to influence best policy in 
   the case of equal confidence
+- Rasa Stack HTTP training endpoint at ``POST /jobs``
 
 Changed
 -------